dcp-815 separate class metadata/checksums.py
