+++ exporter/scratch.py
@@ -165,23 +165,3 @@ class MetadataService:
         return [MetadataResource.from_dict(m) for m in metadata_resources]
 
 
-@dataclass
-class FileChecksums:
-    sha256: str
-    crc32c: str
-    sha1: str
-    s3_etag: str
-
-    @staticmethod
-    def from_dict(data: Dict) -> 'FileChecksums':
-        try:
-            sha256 = data["sha256"]
-            crc32c = data["crc32c"]
-            sha1 = data["sha1"]
-            s3_etag = data["s3_etag"]
-
-            return FileChecksums(sha256, crc32c, sha1, s3_etag)
-        except (KeyError, TypeError) as e:
-            raise MetadataParseException(e)
-
-
